+++ neo4j/src/main/java/org/apache/metamodel/neo4j/utils/MapColumnTypeHandler.java
@@ -1,0 +1,39 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.metamodel.neo4j.utils;
+
+import org.apache.metamodel.schema.ColumnType;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+public class MapColumnTypeHandler extends AbstractColumnTypeHandler {
+    @Override
+    public ColumnType getTypeFromValue(final JSONObject data, final String key) {
+        try {
+            data.getJSONObject(key);
+            return ColumnType.MAP;
+        } catch (final JSONException e) {
+            if (_successor !=  null) {
+                return _successor.getTypeFromValue(data, key);
+            } else {
+                return ColumnType.STRING;
+            }
+        }
+    }
+}
